feat: translate English terms to Spanish across Filament panels
- PaymentResource: navigation label, group, form options (method, status), placeholders
- PatientResource: doctor label, allergy/severity labels, action buttons
- AppointmentResource: procedure label, status/type options
- BudgetResource: status options, odontogram label, notes placeholder, source column
- ProcedurePriceResource: navigation group
- InventoryResource: navigation group
- Roles: navigation group
- CalendarWidget: notification titles and bodies
- PatientAppointmentsChart: heading and dataset label
- PatientStatsOverview: stat labels and descriptions
- FullCalendar: add Spanish locale (locale: 'es')

+++ app/Livewire/PatientPortal/BookAppointment.php
@@ -114,7 +114,7 @@ class BookAppointment extends Component
         $this->validate();
 
         $procedure = ProcedurePrice::findOrFail($this->selectedProcedureId);
-        $duration = $procedure->duration ?? 30;
+        $duration = (int) ($procedure->duration ?? 30);
 
         $date = Carbon::parse($this->selectedDate);
         $timeParts = explode(':', $this->selectedTimeSlot);
